Move Snapshot capabilities to EE (#95)
