+++ .github/workflows/go_lint.yaml
@@ -1,5 +1,8 @@
 name: Go linter for images
 
+env:
+  GO_BUILD_TAGS: "ce se seplus ee csepro"
+
 on:
   pull_request:
   push:
@@ -18,26 +21,27 @@ jobs:
       - name: Setup Go environment
         uses: actions/setup-go@v5
         with:
-          go-version: '1.22'
+          go-version: '1.23.5'
 
       - name: Install golangci-lint
-        run: go install github.com/golangci/golangci-lint/cmd/golangci-lint@v1.59.1
+        run: go install github.com/golangci/golangci-lint/cmd/golangci-lint@v1.61.0
 
       - name: Run Go lint
         run: |
           basedir=$(pwd)
           failed='false'
-          for dir in $(find images -type d); do
-            if ls $dir/go.mod &> /dev/null; then
-              echo "Running linter in $dir"
-              cd $dir
-              golangci-lint run
+          for i in $(find images -type f -name go.mod);do
+            dir=$(echo $i | sed 's/go.mod$//')
+            cd $basedir/$dir
+            # check all editions
+            for edition in $GO_BUILD_TAGS ;do
+              echo "Running linter in $dir (edition: $edition)"
+              golangci-lint run --build-tags $edition
               if [ $? -ne 0 ]; then
-                echo "Linter failed in $dir"
+                echo "Linter failed in $dir (edition: $edition)"
                 failed='true'
               fi
-            cd $basedir
-            fi
+            done
           done
           if [ $failed == 'true' ]; then
             exit 1
