+++ .werf/bundle.yaml
@@ -2,7 +2,7 @@
 ---
 image: bundle
 from: registry.deckhouse.io/base_images/scratch@sha256:b054705fcc9f2205777d80a558d920c0b4209efdc3163c22b5bfcb5dda1db5fc
-fromCacheVersion: "2023-11-27.1"
+fromCacheVersion: "2025-02-18.1"
 import:
 # Rendering .werf/images-digests.yaml is required!
 - image: images-digests
@@ -13,6 +13,11 @@ import:
 - image: python-dependencies
   add: /lib/python/dist
   to: /lib/python/dist
+  after: setup
+# Rendering .werf/choose-edition.yaml is required!
+- image: choose-edition
+  add: /openapi
+  to: /openapi
   after: setup
 git:
 - add: /
@@ -25,7 +30,6 @@ git:
   - enabled
   - hooks
   - monitoring
-  - openapi
   - templates
   - Chart.yaml
   - module.yaml
