+++ images/sds-local-volume-csi/src/go.mod
@@ -5,7 +5,8 @@ go 1.23.4
 require (
 	github.com/container-storage-interface/spec v1.10.0
 	github.com/deckhouse/sds-local-volume/api v0.0.0-20250114155747-5d75d401a787
-	github.com/deckhouse/sds-node-configurator/api v0.0.0-20250114161813-c1a8b09cd47d
+	github.com/deckhouse/sds-local-volume/lib/go/common v0.0.0-20250215220933-9c8f9f5ab53d
+	github.com/deckhouse/sds-node-configurator/api v0.0.0-20250213122438-37fad98a996a
 	github.com/go-logr/logr v1.4.2
 	github.com/golang/protobuf v1.5.4
 	github.com/google/uuid v1.6.0
@@ -71,3 +72,5 @@ require (
 replace github.com/imdario/mergo => github.com/imdario/mergo v0.3.16
 
 replace github.com/deckhouse/sds-local-volume/api => ../../../api
+
+replace github.com/deckhouse/sds-local-volume/lib/go/common => ../../../lib/go/common
