+++ lib/go/common/go.mod
@@ -1,0 +1,3 @@
+module github.com/deckhouse/sds-local-volume/lib/go/common
+
+go 1.23.4
