+++ lib/go/common/pkg/feature/const_ee.go
@@ -1,0 +1,11 @@
+//go:build ee
+
+/*
+Copyright 2025 Flant JSC
+Licensed under the Deckhouse Platform Enterprise Edition (EE) license.
+See https://github.com/deckhouse/deckhouse/blob/main/ee/LICENSE
+*/
+
+package feature
+
+const snapshotsEnabled = true
