+++ lib/go/common/pkg/feature/feature.go
@@ -1,0 +1,5 @@
+package feature
+
+func SnapshotsEnabled() bool {
+	return snapshotsEnabled
+}
